+++ student-management/src/main/java/vn/edu/ptit/PhanHoangAnh/student_management/controller/SubjectController.java
@@ -26,25 +26,25 @@ public class SubjectController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse<Subject>> findSubjectById(@PathVariable int id) {
+    public ResponseEntity<ApiResponse<Subject>> findSubjectById(@PathVariable Long id) {
         Subject subject = this.subjectService.findSubjectById(id);
         return ApiResponse.success(subject);
     }
 
     @PostMapping("/{TeacherId}")
-    public ResponseEntity<ApiResponse<Subject>> saveSubject(@PathVariable int id, @RequestBody Subject subject) {
+    public ResponseEntity<ApiResponse<Subject>> saveSubject(@PathVariable Long id, @RequestBody Subject subject) {
         Subject subjectSave = this.subjectService.saveSubject(id, subject);
         return ApiResponse.created(subjectSave);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ApiResponse<Subject>> updateSubjectById(@PathVariable int id, @RequestBody Subject subject) {
+    public ResponseEntity<ApiResponse<Subject>> updateSubjectById(@PathVariable Long id, @RequestBody Subject subject) {
         Subject subjectUpdate = this.subjectService.updateSubjectById(id, subject);
         return ApiResponse.success(subjectUpdate);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<String>> deleteSubjectById(@PathVariable int id) {
+    public ResponseEntity<ApiResponse<String>> deleteSubjectById(@PathVariable Long id) {
         this.subjectService.deleteSubjectById(id);
         return ApiResponse.success("delete success...");
     }
