+++ student-management/src/main/java/vn/edu/ptit/PhanHoangAnh/student_management/service/ClassLogbookService.java
@@ -4,9 +4,9 @@ import vn.edu.ptit.PhanHoangAnh.student_management.entity.ClassLogbook;
 import java.util.List;
 
 public interface ClassLogbookService {
-    ClassLogbook findClassLogbookById(int id);
+    ClassLogbook findClassLogbookById(Long id);
     List<ClassLogbook> findAllClassLogbook();
-    ClassLogbook saveClassLogbook(int classId, ClassLogbook classLogbook);
-    ClassLogbook updateClassLogbookById(int id, ClassLogbook classLogbook);
-    void deleteClassLogbookById(int id);
+    ClassLogbook saveClassLogbook(Long classId, ClassLogbook classLogbook);
+    ClassLogbook updateClassLogbookById(Long id, ClassLogbook classLogbook);
+    void deleteClassLogbookById(Long id);
 }
